+++ src/export.rs
@@ -300,7 +300,7 @@ mod export_tests {
         let tutorial_text = r#"Title: Lottie Tutorial
 Credit: Written by
 Author: René Coignard
-Draft date: Version 0.2.11
+Draft date: Version 0.2.12
 Contact:
 contact@renecoignard.com